Extract class MiningWorker
