+++ src/main/app/MiningFramework.groovy
@@ -8,6 +8,8 @@ import java.io.File
 import java.util.ArrayList
 import java.text.SimpleDateFormat
 import static groovy.io.FileType.DIRECTORIES
+import java.util.concurrent.BlockingQueue
+import java.util.concurrent.LinkedBlockingQueue
 
 import main.arguments.*
 import main.project.*
@@ -17,6 +19,7 @@ import main.exception.UnstagedChangesException
 import main.exception.UnexpectedPostScriptException
 import main.exception.NoAccessKeyException
 import main.util.*
+
 
 class MiningFramework {
 
@@ -80,30 +83,15 @@ class MiningFramework {
     public void start() {
         projectList = processProjects(projectList)
 
-        for (project in projectList) {
-            printProjectInformation(project)
-            
-            if (project.isRemote()) {
-                cloneRepository(project, LOCAL_PROJECT_PATH)
-            } else {
-                checkForUnstagedChanges(project);
-            }
-            
-            List<MergeCommit> mergeCommits = project.getMergeCommits(arguments.getSinceDate(), arguments.getUntilDate()) // Since date and until date as arguments (dd/mm/yyyy).
-            for (mergeCommit in mergeCommits) {
-                if (applyFilter(project, mergeCommit)) {
-                    printMergeCommitInformation(mergeCommit)
-                    collectStatistics(project, mergeCommit)
-                    collectExperimentalData(project, mergeCommit)
-                }
-            }
-
-            if(arguments.isPushCommandActive()) // Will push.
-                pushResults(project, arguments.getResultsRemoteRepositoryURL())
-            
-            endProjectAnalysis()
+        // TODO: Change logic to move projects to BlockingQueue
+        BlockingQueue<Project> projectQueue = new LinkedBlockingQueue<Project>()
+        for (Project project : projectList ) {    
+            projectQueue.add(project)
         }
 
+        def worker = new MiningWorker(dataCollector, statCollector, commitFilter, projectQueue, LOCAL_PROJECT_PATH)
+
+        worker.run()
         processOutput()
     }
 
@@ -122,26 +110,6 @@ class MiningFramework {
         }
     }
 
-    private void checkForUnstagedChanges(Project project) {
-        String gitDiffOutput = ProcessRunner.runProcess(project.getPath(), "git", "diff").getText()
-
-        if (gitDiffOutput.length() != 0) {
-            throw new UnstagedChangesException(project.getName())
-        }
-    }
-
-    private boolean applyFilter(Project project, MergeCommit mergeCommit) {
-        return commitFilter.applyFilter(project, mergeCommit)
-    }
-
-    private void collectStatistics(Project project, MergeCommit mergeCommit) {
-        statCollector.collectStatistics(project, mergeCommit)
-    }
-
-    private void collectExperimentalData(Project project, MergeCommit mergeCommit) {
-        dataCollector.collectExperimentalData(project, mergeCommit)
-    }
-
     private ArrayList<Project> processProjects(ArrayList<Project> projects) {
         return projectProcessor.processProjects(projects)
     }
@@ -150,78 +118,8 @@ class MiningFramework {
         outputProcessor.processOutput()
     }
 
-    private void pushResults(Project project, String remoteRepositoryURL) {
-        Project resultsRepository = new Project('', remoteRepositoryURL)
-        printPushInformation(remoteRepositoryURL)
-        String targetPath = "${LOCAL_RESULTS_REPOSITORY_PATH}/resultsRepository"
-        cloneRepository(resultsRepository, targetPath)
-
-        // Copy output files, add, commit and then push.
-        FileManager.copyDirectory(getOutputPath(), "${targetPath}/output-${project.getName()}")
-        Process gitAdd = ProcessRunner.runProcess(targetPath, 'git', 'add', '.')
-        gitAdd.waitFor()
-
-        def nowDate = new Date()
-        def sdf = new SimpleDateFormat("dd/MM/yyyy")
-        Process gitCommit = ProcessRunner
-            .runProcess(targetPath, 'git', 'commit', '-m', "Analysed project ${project.getName()} - ${sdf.format(nowDate)}")
-        gitCommit.waitFor()
-        gitCommit.getInputStream().eachLine {
-            println it
-        }
-
-        Process gitPush = ProcessRunner.runProcess(targetPath, 'git', 'push', '--force-with-lease')
-        gitPush.waitFor()
-        gitPush.getInputStream().eachLine {
-            println it
-        }
-
-        FileManager.delete(new File(targetPath))
-    }
-
-    private void cloneRepository(Project project, String target) {
-        
-        println "Cloning repository ${project.getName()} into ${target}"
-
-        File projectDirectory = new File(target)
-        if(projectDirectory.exists()) {
-            FileManager.delete(projectDirectory)
-        }
-        projectDirectory.mkdirs()
-        
-        String url = project.getPath()
-
-        if (arguments.providedAccessKey()) {
-            String token = arguments.getAccessKey();
-            String[] projectOwnerAndName = project.getOwnerAndName()
-            url = "https://${token}@github.com/${projectOwnerAndName[0]}/${projectOwnerAndName[1]}"
-        }
-
-        ProcessBuilder builder = ProcessRunner.buildProcess('./', 'git', 'clone', url, target)
-        builder.redirectOutput(ProcessBuilder.Redirect.INHERIT)
-    
-        Process process = ProcessRunner.startProcess(builder)
-        process.waitFor()
-  
-        project.setPath(target)
-    }
-
-    private void printProjectInformation(Project project) {
-        println "PROJECT: ${project.getName()}"
-    }
-
-    private void printMergeCommitInformation(MergeCommit mergeCommit) {
-        println "Merge commit: ${mergeCommit.getSHA()}"
-    }
-
     private void printPushInformation(String url) {
         println "Proceeding to push output files to ${url}."
-    }
-
-    private void endProjectAnalysis() {
-        File projectDirectory = new File(LOCAL_PROJECT_PATH)
-        if (projectDirectory.exists())
-            FileManager.delete(new File(LOCAL_PROJECT_PATH))
     }
 
     public void setProjectList(ArrayList<Project> projectList) {
